该版本能在液晶显示器上实时显示温度并且启动了内存使其不闪烁，定义了按键LED1当按键被按下时LED1状态翻转，BSP_DS18B20_ReadByte()函数的data不初始化是就会读取温度出错???

+++ Temp_F407_OS/User/BSP/bsp_18b20/bsp_18b20.h
@@ -78,7 +78,7 @@
 
 CPU_BOOLEAN  BSP_18B20_Init(void);
 CPU_BOOLEAN  BSP_18B20_GetTemp(uint16_t *temp);
-uint32_t     BSP_18B20_GetTempFast(void);
+uint16_t     BSP_18B20_GetTempFast(void);
 float        BSP_18B20_TempTran(uint16_t temp);
